adicionado mais atributos ao metodo PUT de Voluntario

+++ src/main/java/com/solidarity/solidarity_backend/services/validation/VoluntarioUpdateValidator.java
@@ -33,7 +33,7 @@ public class VoluntarioUpdateValidator implements ConstraintValidator<Voluntario
 
         @SuppressWarnings("unchecked")
         Map<String, String> map = (Map<String, String>)request.getAttribute(HandlerMapping.URI_TEMPLATE_VARIABLES_ATTRIBUTE);
-        Integer uriId = Integer.parseInt(map.get("id"));
+        Long uriId = Long.parseLong(map.get("id"));
 
         List<FieldMessage> list = new ArrayList<>();
